consolidated test-scenario-runner to single module
